requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.2.0
xlrd>=2.0.0
xlwt>=1.3.0
xlutils>=2.0.0
python-dotenv>=1.0.0
fastexcel>=0.16.0
python-calamine>=0.4.0
pywin32>=306

